+++ Beagle/UIImage+Resize.m
@@ -129,7 +129,7 @@
                                                8,
                                                0,
                                                colorSpace,
-                                               kCGImageAlphaPremultipliedLast);
+                                               (int)kCGImageAlphaPremultipliedLast);
     CGColorSpaceRelease(colorSpace);
 	
     // Rotate and/or flip the image if required by its orientation
